made create exercise a pop up
