+++ app/src/components/Exercise/AnswerField.tsx
@@ -20,7 +20,7 @@ export const AnswerField = ({index, answerObject, callback}:Inputs) => {
             {/** Correct answer button */}
             <div className="form-control w-1/5">
                 <label className="label">
-                    <span className="label-text">Right answer</span>
+                    <span className="label-text">Resposta correta</span> {/**Right answer */}
                 </label>    
                     <input type="checkbox" className="" defaultChecked={answerObject.isCorrect} onChange={()=>{answerObject.setIsCorrect(!answerObject.isCorrect); callback(index, answerObject)}}/> 
             </div>
@@ -28,13 +28,14 @@ export const AnswerField = ({index, answerObject, callback}:Inputs) => {
             {/** Answer text field */}
             <div className="form-control w-2/5">
                 <label className="label">
-                    <span className="label-text">Answer</span>
+                    <span className="label-text">Resposta</span> {/**Resposta */}
                 </label>
                 <textarea
                     className="input input-bordered w-full h-16"   
-                    placeholder="Add a possible answer to your exercise"
+                    placeholder="Adicione uma resposta possível ao seu exercício" 
                     onChange={(e) => {answerObject.setText(e.target.value); callback(index, answerObject)}}
-                />  
+                    
+                />  {/**Add a possible answer to your exercise */}
             </div>
 
             {/** Feedback input field*/}
@@ -44,12 +45,12 @@ export const AnswerField = ({index, answerObject, callback}:Inputs) => {
                 </label>
                 <textarea
                     className="input input-bordered w-full h-16"
-                    placeholder="Add feedback to your exercise"
+                    placeholder="Adicione feedback ao seu exercício"
                     onChange={(e) => {
                             answerObject.setFeedback(e.target.value); 
                             callback(index, answerObject);
                         }}    
-                />
+                />  {/**Add feedback to your exercise */}
             </div>
 
         </div>
