+++ app/src/pages/SectionEdit.tsx
@@ -10,11 +10,14 @@ import { toast } from 'react-toastify';
 import SectionServices from '../services/section.services';
 import ExerciseServices from '../services/exercise.services';
 import { CreateLecture } from '../components/LecturePage';
+import { CreateExercise } from '../components/Exercise/CreateExercisePopUp';
 
 // Components
 import Loading from './Loading'
 import Layout from '../components/Layout'
 import { ExerciseArea } from "../components/ExerciseArea";
+import { AnswerField } from "../components/Exercise/AnswerField";
+
 
 // Interface
 import { type Section } from '../interfaces/CourseDetail'
@@ -24,7 +27,6 @@ import { Answer } from "../interfaces/Answer";
 // Icons
 import ArrowLeftIcon from '@heroicons/react/24/outline/ArrowLeftIcon'
 import { PlusIcon } from "@heroicons/react/24/outline";
-import { AnswerField } from "../components/Exercise/AnswerField";
 
 // Backend URL from .env file (automatically injected) 
 const backendUrl = import.meta.env.VITE_BACKEND_URL;
@@ -39,11 +41,6 @@ type SectionPartial = {
   description: string;
 };
 
-type ExercisePartial = {
-  title: string;
-  question: string;
-  answers: Answer[];
-};
 
 /**
  * SectionEdit component
@@ -52,9 +49,6 @@ type ExercisePartial = {
  */
 const SectionEdit = () => {
   const { cid, sid } = useParams();
-  const [tempAnswers, setTempAnswer] = useState<Answer[]>([new Answer(), new Answer()]);
-  const [answerFieldIndex, setAnswerFieldIndex] = useState(tempAnswers.length);
-  const [answerField, setAnswerField] = useState<JSX.Element[]>([<AnswerField index={0} answerObject={tempAnswers[0]}  callback={updateAnswer}/>, <AnswerField index={1} answerObject={tempAnswers[1]}  callback={updateAnswer}/>]);
 
   const token = "dummyToken";
   //const token = useToken()
@@ -62,6 +56,7 @@ const SectionEdit = () => {
   // Component state
   const [section, setSection] = useState<Section>();
   const [exercises, setExercises] = useState<Exercise[]>([]);
+
 
   // Fetch the section data from the server.
   const { data: sectionData, error: sectionError } = useSWR(
@@ -71,7 +66,7 @@ const SectionEdit = () => {
 
   console.log(sectionData?.parentCourse);
   if (sectionData != undefined && sectionData.parentCourse != cid) {
-    toast.error("Parent course doesn't match section.");
+    toast.error("Parent course doesn't match section."+ sectionData.parentCourse + " " + cid);
   }
 
   // Fetch the exercises data from the server.
@@ -80,19 +75,13 @@ const SectionEdit = () => {
     ExerciseServices.getExerciseDetail
   );
 
-
-
   // Create Form Hooks
   const {
     register: registerSection,
     handleSubmit: handleSectionUpdate,
     formState: { errors: sectionErrors },
   } = useForm<Inputs>();
-  const {
-    register: registerExercise,
-    handleSubmit: handleExerciseAdd,
-    formState: { errors: exerciseErrors },
-  } = useForm<ExercisePartial>();
+
 
   /**
    * SubmitHandler: update section
@@ -111,70 +100,14 @@ const SectionEdit = () => {
       .catch((err) => toast.error(err));
   };
 
-  /**
-   * SubmitHandler: add exercise
-   * @param exerciseData  The exercise data to be added
-   * @returns     void
-   */
-  const onExerciseAdd: SubmitHandler<ExercisePartial> = (exerciseData) => {
-    const exerciseToAdd: ExercisePartial = {
-      title: exerciseData.title,
-      question: exerciseData.question,
-      answers: exerciseData.answers,
-    };
 
-    ExerciseServices.addExercise(exerciseToAdd, /*token,*/ sid)
-      .then((res) => {
-        toast.success("Added exercise");
-        window.location.reload();
-      })
-      .catch((err) => toast.error(err));
-  };
 
   // Render onError and onLoading
   if (sectionError) return <p>"An error has occurred."</p>;
   if (!sectionData || !exerciseData) return <Loading />;
-
-  /**
-   * Add new answer 
-   * @param index index of the answer to be added
-   * @returns 
-   */
-  function addOrGetAnswer(index: number) {
-    if (tempAnswers[index] == undefined) {
-      tempAnswers[index] = new Answer();
-    }
-    return tempAnswers[index];
-  }
-
-  /** 
-   * Update answer to match inputted data
-   */ 
-  function updateAnswer(answerFieldIndex: number, answer: Answer) {
-    tempAnswers[answerFieldIndex] = answer;
-    setTempAnswer(tempAnswers);
-  }
-
-  /**
-   * Add new answer field
-   * 
-   */
-  function addAnswerField() {
-    setAnswerFieldIndex(answerFieldIndex + 1);
-    setAnswerField([
-      ...answerField,
-      <AnswerField
-        index={answerFieldIndex}
-        answerObject={addOrGetAnswer(answerFieldIndex)}
-        callback={updateAnswer}
-      />,
-    ]);
-  }
-
  
   return (
     <Layout meta="Section edit page">
-      <div className="w-full" onScroll={(e) => {addAnswerField();addAnswerField();}}>
         {/** Course navigation */}
         <div className="navbar bg-base-100">
           <div className="flex-1">
@@ -199,7 +132,7 @@ const SectionEdit = () => {
           >
             {/** Section Title Field */}
             <div className="flex flex-col space-y-2">
-              <label htmlFor="title">Título</label>
+              <label htmlFor="title">Título</label> {/*Titel */}
               <input
                 type="text"
                 defaultValue={section?.title || sectionData?.title}
@@ -212,7 +145,7 @@ const SectionEdit = () => {
 
             {/** Section Description Field */}
             <div className="flex flex-col space-y-2">
-              <label htmlFor="description">Descrição</label>
+              <label htmlFor="description">Descrição</label> {/**Description */}
               <textarea
                 rows={4}
                 defaultValue={section?.description || sectionData?.description}
@@ -221,17 +154,14 @@ const SectionEdit = () => {
                 {...registerSection("description", { required: false })}
               />
               {sectionErrors.description && (
-                <span>Este campo é obrigatório!</span>
+                <span>Este campo é obrigatório!</span> /**This field is obligator */
               )}
             </div>
-
-            
 
             <button type="submit" className="std-button ml-auto">
               Atualizar
             </button>
-            
-
+          
           </form>
               
 
@@ -261,65 +191,21 @@ const SectionEdit = () => {
           {/** New exercise area */}
 
           <div className="flex flex-col w-full mb-4">
-            <span className="text-xl font-medium">Add new exercise</span>
+            <span className="text-xl font-medium">Adicionar novo exercício.</span> {/* Add new exercise*/}
+          </div>
+          
+          <div className="divider"></div>
+          
+          <div className="navbar bg-none p-6">
+              <div className="flex-1">
+                  
+              {/** Create new lecture */}
+              <CreateExercise sid={sid} cid={cid}/>
+              </div>
           </div>
 
-          <form
-            onSubmit={handleExerciseAdd(onExerciseAdd)}
-            className="flex flex-col justify-content align-items space-evenly w-full space-y-2"
-          >
-            {/** Exercise Title Field */}
-            <div className="form-control w-full" >
-              <label className="label">
-                <span className="label-text">Title</span>
-              </label>
-              <input
-                type="text"
-                placeholder="Some awesome title"
-                className="input input-bordered w-full"
-                {...registerExercise("title", { required: true })}
-              />
-              {exerciseErrors.title && <span>This field is required</span>}
-            </div>
-             
-            {/** Exercise Question Field */}
-            <div className="form-control w-full">
-              <label className="label">
-                <span className="label-text">Question</span>
-              </label>
-              <textarea 
-                className="textarea textarea-bordered h-24"
-                placeholder="Add a question to your exercise"
-                {...registerExercise("question", { required: true })}
-              />
-            </div>
 
-            {/** Exercise Answers Field */}
-            {answerField}
-
-            <div className="flex justify-between items-center border rounded p-1">
-              <div>
-                <button
-                  className="btn" type="button"
-                  onClick={() => {
-                    if (answerFieldIndex < 4) {
-                      addAnswerField();
-                    } else {
-                      toast.error("Maximum number of answers reached.");
-                    }
-                  }}
-                >
-                  <PlusIcon width={24} />
-                </button>
-              </div>
-            </div>
-
-            <button type="submit" className="std-button ml-auto" {...registerExercise('answers', {value: tempAnswers})}>
-              Add Exercise
-            </button>
-          </form>
         </div>
-      </div>
     </Layout>
   );
 };
